backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "slowapi>=0.1.9",
    "uvicorn>=0.44.0",
]
